fix: updated react-datepicker, fixed bugs and code refactoring
