+++ packages/datepicker/package.json
@@ -39,13 +39,11 @@
   "dependencies": {
     "@equinor/eds-icons": "^0.6.2",
     "@equinor/fusion-react-styles": "^0.2.0",
-    "@equinor/fusion-wc-icon": "^1.0.2",
-    "@equinor/fusion-wc-popover": "^1.0.1",
-    "react-datepicker": "^3.8.0"
+    "react-datepicker": "^4.1.1"
   },
   "devDependencies": {
-    "@types/react-datepicker": "^3.1.8",
+    "@types/react-datepicker": "^4.1.3",
     "jss-cli": "^6.0.2",
-    "typescript": "^4.1.3"
+    "typescript": "^4.3.5"
   }
 }
